pyproject.toml
[project]
name = "agno-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.3.1",
    "duckduckgo-search>=8.0.0",
    "exa-py>=1.12.0",
    "groq>=0.22.0",
    "openai>=1.74.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.55",
]
